DE2/ingestion/requirements.txt
requests>=2.31
pyarrow>=15.0
google-cloud-storage>=2.14
